--- src/main/java/_9oormthonuniv/springjwt/controller/MainController.java
@@ -1,16 +1,0 @@
-package _9oormthonuniv.springjwt.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-@ResponseBody
-public class MainController {
-
-    @GetMapping("/")
-
-    public String mainP() {
-        return "Main Controller";
-    }
-}
