+++ src/main/java/_9oormthonuniv/springjwt/domain/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package _9oormthonuniv.springjwt.repository;
+package _9oormthonuniv.springjwt.domain.post.repository;
 
-import _9oormthonuniv.springjwt.entity.PostEntity;
+import _9oormthonuniv.springjwt.domain.post.entity.PostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<PostEntity, Long> {
