+++ src/main/java/_9oormthonuniv/springjwt/domain/user/controller/AdminController.java
@@ -1,4 +1,4 @@
-package _9oormthonuniv.springjwt.controller;
+package _9oormthonuniv.springjwt.domain.user.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
