+++ src/main/java/_9oormthonuniv/springjwt/global/config/AwsS3Config.java
@@ -1,4 +1,4 @@
-package _9oormthonuniv.springjwt.config;
+package _9oormthonuniv.springjwt.global.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
