+++ src/main/java/_9oormthonuniv/springjwt/global/config/SecurityConfig.java
@@ -1,6 +1,6 @@
-package _9oormthonuniv.springjwt.config;
+package _9oormthonuniv.springjwt.global.config;
 
-import _9oormthonuniv.springjwt.jwt.LoginFilter;
+import _9oormthonuniv.springjwt.global.jwt.LoginFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -8,7 +8,6 @@ import org.springframework.security.config.annotation.authentication.configurati
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
